drivers - boot: [FIX] rescan PCI buses in two phases
The NFB endpoint pairing procedure is activated upon the primary PCIe EP
initialization. It searches through all available secondary PCIe EPs.
If the secondary PCIe is not (yet) in the system, it will be not paired,
which is unwanted behaviour.

The PCI rescan procedure initializes all PCIe devices found on specified
bridge. On some system configuration (e.g. when using bifurcation),
the secondary PCIe EP shares a bridge with the primary PCIe EP.

In such case, the driver must ensure, that a PCI rescan of that shared
bridge is done at last, in order to activate the pairing procedure
at the right time (after all secondary PCIe EPs are initialized).

Linux kernel ensures, that the secondary PCIe EPs on the same bridge
as the primary PCIe EP are already available during the primary PCIe EP
probe (which also triggers the NFB endpoint pairing procedure).