nfb-xvc: [FEATURE] use libpci instead of driver
The nfb-xvc now supports driverless access through the libpci library.
Minimal version of the libpci is 3.6.3; for older systems (like RHEL7)
is used the original mechanism.

For libpci access the device parameter -d must be in PCI BDF format.

The xvc kernel driver is disabled and doesn't built by default.
For usage of nfb-xvc on older systems must be therefore manually enabled.