libnfb - open: [FIX] check whole path string
In the short path feature (for example "nfb-info -d 1" for the "/dev/nfb1")
we must check if the device path string was completely parsed.
User can specify the device in other manner (e.g. "-d 0000:17:00.0"),
which can be acceptable form for some functions.

But in this case the sscanf has successfully parsed just "0000" as a valid number
and nfb_open used it in snprintf("/dev/nfb%d", nr), which can lead to unwanted device.