libnfb - nfb_comp_lock: [FIX] check errno instead of the ioctl return value
The ioctl return value is always -1 for unsuccessful request.
Therefore particular error must be obtained by errno.

This commit fixes the nfb_comp_lock behaviour: now the while loop retries to
lock endlessly - acting similar as spinlocks. This behaviour is intended for
easy use in the userspace code (there is currently a limit of 100 retries /
100ms in the kernel driver, ensuring some deadlock prevention).

However this behaviour is not suitable for all cases, which will be
covered better by try-lock-like mechanism.