libnetcope - dma_ctrl_ndp [FIX]: use nfb_comp_trylock instead of nfb_comp_lock
The nc_ndp_ctrl_start locks the firmware component for exclusive
access. This lock is however not intended to be held for a short time,
so the usage of basic nfb_comp_lock is improper.

Instead of potentially endless lock loop is better to use
nfb_comp_trylock and check for -EBUSY return value immediately and only once.