feat(ndp-calypte): add flipping of valid flag
The Valid Flag present in the DMA header within DMA Calypte module was
previously set to 1 every time a new packet has arrived which required
the clearing of the flags of packets that were already read. This change
introduces flag flipping which considers both 0 and 1 to be valid values
based of which run through a buffer is considered. Upon starting of a
channel, the valid value is 1 whereas upon header pointer rollover, this
value flips to 0 which is then considered as a valid value for a second
run through the buffer. Upon next pointer rollover, the valid value
flips back to 1 and so on. Because of that, the header buffer is mapped
as READ_ONLY since hardware performs flag flip on its own.