feat(dma-registers): add EXPER register for Peer-to-Peer transfers
This register has been introduced to DMA Calypte for experimental
features and now it contains only one functional bit (bit 0) which
enables a Peer-to-peer transfer for the specific DMA channel. This
register is cleared when the communication is done between a software
and a DMA controller since enabling Peer-to-peer introduces custom
pointer translation which whould lead to a malformation of data.